Calculate Energy for Activity

Computes the energy requirement for activity (MJ/head/day). This approach follows the IPCC Tier 2 energy partitioning method and applies the following general equation: energy_activity = cact * energy_maintenance (except for sheep: cact * average_weight) where cact is an activity coefficient reflecting the animal’s feeding and management situation. (MJ d^-1 kg^-1).

calc_net_energy_activity(
  animal,
  cohort,
  nemain,
  average_weight,
  activity_fraction,
  high_activity_fraction
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CHK: chickens

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

cohort

Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:

FA: adult females (from age at first parturition)

FS: sub-adult females (from weaning to age at first parturition)

FJ: juvenile females (from birth to weaning)

MA: adult males (from age at first breeding)

MS: sub-adult males (from weaning to age at first breeding)

MJ: juvenile males (from birth to weaning)

nemain

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

average_weight

Numeric. Average live weight over the cohort stage. Computed by accounting for the share of offtaken animals within the cohort, using their slaughter weight, and the potential final weight of animals that remain in the cohort (kg).

activity_fraction

Numeric. Proportion of the assessment period during which the animal performs low-intensity movement typical of stall-feeding or near-field grazing, characterized by minimal walking distances and flat terrain (fraction).

high_activity_fraction

Numeric. Proportion of the assessment period during which the animal engages in sustained locomotion associated with herding or long-distance grazing, typically involving extended walking distances and/or uneven or hilly terrain (fraction).

Value

Numeric. Energy required for activity, defined as the amount of energy needed to obtain food, water and shelter (example stall, grazing large areas). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

Details

The activity coefficient cact reflects the animal’s feeding and management conditions. For CTL, BFL, SHP, and GTS, cact is calculated as a weighted average of the different activity levels over the assessment period. Reference coefficients are derived from NRC (1996) and AFRC (1993) for CTL, BFL, SHP, and GTS; NRC (1998) for PGS; and Wardeh (2004) for CML.

References

NRC (1998). Nutrient Requirements of Swine, 10th Revised Edition. National Academies Press, Washington, DC.

NRC (1996). Nutrient Requirements of Beef Cattle, 7th Revised Edition. National Academies Press, Washington, DC.

AFRC (1993). Energy and Protein Requirements of Ruminants. CAB International, Wallingford, UK.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.4; Table 10.5.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.4; Table 10.5.

Wardeh, M. F. (2004). The nutrient requirements of the dromedary camel. Journal of Camel Science. 2004;1:37–45. The Camel Applied Research and Development Network (CARDN), Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).