Calculate Energy for Lactation

Computes the energy requirement for lactation (MJ/head/day), defined as the energy needed to support milk production by lactating females. This approach follows the IPCC Tier 2 partitioning method and applies species-specific equations for lactation energy requirements.

calc_net_energy_lactation(
  animal,
  cohort,
  milking_fraction,
  milk_yield,
  milk_fat,
  idle,
  gest,
  litsize,
  dr1,
  ckg,
  wkg,
  lact,
  parturition_rate
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CHK: chickens

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

cohort

Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:

FA: adult females (from age at first parturition)

FS: sub-adult females (from weaning to age at first parturition)

FJ: juvenile females (from birth to weaning)

MA: adult males (from age at first breeding)

MS: sub-adult males (from weaning to age at first breeding)

MJ: juvenile males (from birth to weaning)

milking_fraction

Numeric. Share of adult females lactating within the assessment duration. Applies to species = CML, CTL, BFL, SHP, GTS. (fraction).

milk_yield

Numeric. Average milk yield per milk-producing animal during the assessment duration (kg/head/day). This value can be calculated by dividing the total milk destinated to human consumption produced per milk-producing animal over the assessment duration by the length of the assessment period.

milk_fat

Numeric. Milk fat fraction (kg fat / kg milk).

idle

Numeric. Period during which the animal is not performing any productive physiological function such as pregnancy or lactation (days).

gest

Numeric. Duration of pregnancy period (days).

litsize

Numeric. Average number of offspring born per parturition (#). This value can be calculated as the total number of offspring born divided by the total number of parturitions during the year.

dr1

Numeric. Percentage of deaths in a herd over a year for juvenile cohorts (i.e. FJ and MJ) (fraction).

ckg

Numeric. Live body weight of the animal at birth (kg).

wkg

Numeric. Live body weight of the animal at weaning (kg).

lact

Numeric. Duration of the lactation period, defined as the number of days during which the animal is lactating (days).

parturition_rate

Numeric. Average annual number of parturition per female animal (#). A herd-level reproductive performance indicator calculated as the total number of parturitions (born) occurring during a year divided by the number of adult females potentially able to give birth during that year.

Value

Numeric. Energy required for lactation. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

Details

Energy for lactation (energy_lactation) represents the additional energy required to support milk synthesis during lactation. Following IPCC Tier 2 guidelines, lactation energy is calculated as a function of the quantity of milk produced and a species-specific energy cost per unit of milk.

For CTL, BFL, CML, SHP and GTS:

Total milk production includes:

milk extracted for human consumption (milk_yield)

milk consumed directly by offspring (milk_for_offspring)

Lactation energy requirements are applied only to adult females and are scaled by the proportion of milked (milking_fraction) or lactating (parturition_rate) animals within the cohort.

In general form, lactation energy is computed as:

$$ energy\_lactation = (milk\_yield \times milking\_fraction + milk\_for\_offspring) \times energy\_milk$$

where:

energy_milk is a species-specific coefficient representing the net energy cost of producing one kilogram of milk (MJ kg\(^{-1}\))

Species-specific values of energy_milk are:

CTL, BFL: estimated as a function of milk fat content, \(1.47 + 0.40 \times (milk\_fat \times 100)\) (NRC 1989),

CML: 4.063 (Wardeh, 2004),

SHP: 4.6 (AFRC, 1993; AFRC, 1995),

GTS: 3.0 (AFRC, 1998).

milk_for_offspring is the daily amount of milk required to rear offspring across the year (kg day\(^{-1}\)). It is calculated assuming that 5 kg of milk are required for each kilogram of live-weight gain up to weaning, using the equation below.

$$ milk\_for\_offspring = \frac{parturition\_rate \times 5 \times (wkg - ckg)} {365} \code(AFRC, 1990) $$

For SHP and GTS, milk_for_offspring also accounts for the average number of offspring per birth by multiplying by litsize.

For PGS — NRC (1998):

For pigs, lactation energy accounts only for the milk consumed directly by offspring (milk_for_offspring) adjusted by the fraction of the reproductive cycle spent in lactation (cadj), and is calculated as follows:

$$ energy\_lactation = litsize \times (1 - 0.5 \times dr1) \times \left( \frac{0.02059 \times (wkg - ckg) \times 1000}{lact} - \frac{0.3766}{0.67} \right) \times cadj $$

where:

\(0.02059\) is the coefficient for lactation energy requirement (MJ g live-weight\(^{-1}\)),

\(0.3766\) is the coefficient for sow weight loss during lactation (\(C_{wloss}\), MJ head\(^{-1}\) day\(^{-1}\)),

\(0.67\) is the efficiency of conversion of dietary intake to milk energy (\(C_{conv}\), fraction),

litsize is litter size,

dr1 is the proportion of stillborn piglets,

\(wkg\) and \(ckg\) are live weights at weaning and at birth, respectively (kg),

\(cadj\) is the fraction of the reproductive cycle spent in lactation calculated as

$$ cadj = \frac{lact}{idle + gest + lact} $$

References

AFRC (1998) The Nutrition of Goats. Wallingford: CAB International. Animut G., Puchala R., Goetsch A.L., Patra A.K., Sahlu T., Varel V.H., Wells J.

AFRC (1995). Energy and Protein Requirements of Ruminants. CAB International, Wallingford, UK.

AFRC (1993). Energy and Protein Requirements of Ruminants. CAB International, Wallingford, UK.

AFRC (1990). Nutritive Requirements of Ruminant Animals: Energy. Rep. 5. Wallingford, UK: CAB International.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.8-10.10.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.8-10.10.

NRC (1998). Nutrient Requirements of Swine, 10th Revised Edition. National Academies Press, Washington, DC.

NRC (1989) Nutrient Requirements of Dairy Cattle, 6th Ed. . Washington, D.C. U.S.A: National Academy Press.

Wardeh, M. F. (2004). The nutrient requirements of the dromedary camel. Journal of Camel Science. 2004;1:37–45. The Camel Applied Research and Development Network (CARDN), Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).