Calculate Energy for Maintenance

Computes the energy requirement for maintenance (MJ/head/day). This approach follows the IPCC Tier 2 partitioning method, and applies this general equation: energy_maintenance = cmain * average_weight^0.75 where cmain is a category-specific coefficient (MJ d^-1 kg^-1).

calc_net_energy_maintenance(
  animal,
  cohort,
  average_weight,
  milking_fraction = NA_real_,
  offtake_rate = NA_real_,
  afc = NA_real_
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

cohort

Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:

FA: adult females (from age at first parturition)

FS: sub-adult females (from weaning to age at first parturition)

FJ: juvenile females (from birth to weaning)

MA: adult males (from age at first breeding)

MS: sub-adult males (from weaning to age at first breeding)

MJ: juvenile males (from birth to weaning)

average_weight

Numeric. Average live weight over the cohort stage. Computed by accounting for the share of offtaken animals within the cohort, using their slaughter weight, and the potential final weight of animals that remain in the cohort (kg).

milking_fraction

Numeric. Share of adult females lactating within the assessment duration. Applies to species = CML, CTL, BFL, SHP, GTS. (fraction).

offtake_rate

Numeric. Annual proportion of animals removed from the herd for each sex-age cohort (fraction).

afc

Numeric. Age at first parturition for female breeding animals (days)

Value

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

Details

The maintenance coefficient \(cmain\) reflects basal metabolic requirements and differs by species, physiological status, and sex. For selected cohorts it is computed as a weighted average to account for:

lactating vs. non-lactating females (CTL, BFL),

intact vs. castrated males (CTL, BFL, SHP). Offtaken animals are assumed castrated,

animals below vs. above one year of age (SHP).

Specific cofficients by species and cohort:

CTL and BFL (NRC, 1996; AFRC, 1993, 1995):

FA: \(cmain = 0.386 \times milking\_fraction + 0.322 \times (1 - milking\_fraction)\)

FS, FJ, MJ: \(cmain = 0.322\)

MA, MS: \(cmain = 0.322 \times offtake\_rate + 0.370 \times (1 - offtake\_rate)\)

CML (Wardeh, 2004):

All cohorts: \(cmain = 0.435\)

GTS (AFRC, 1993, 1995):

All cohorts: \(cmain = 0.315\)

SHP (AFRC, 1993, 1995):

FA: \(cmain = 0.217\)

FJ: \(cmain = 0.236\)

FS: \(cmain = 0.236 \times (365/afc) + 0.217 \times ((afc - 365)/afc)\)

MA: \(cmain = 0.217 \times offtake\_rate + (0.217 \times 1.15) \times (1 - offtake\_rate)\)

MJ: \(cmain = 0.236 \times offtake\_rate + (0.236 \times 1.15) \times (1 - offtake\_rate)\)

MS: A weighted average of juvenile and adult male coefficients using afc, with an additional intact-male adjustment (multiplied by 1.15) weighted by offtake\_rate.

PGS (NRC, 1998):

All cohorts: $$cmain = 0.4435$$

References

NRC (1998). Nutrient Requirements of Swine, 10th Revised Edition. National Academies Press, Washington, DC.

NRC (1996). Nutrient Requirements of Beef Cattle, 7th Revised Edition. National Academies Press, Washington, DC.

AFRC (1995). Energy and Protein Requirements of Ruminants. CAB International, Wallingford, UK.

AFRC (1993). Energy and Protein Requirements of Ruminants. CAB International, Wallingford, UK.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories. Chapter 10: Emissions from Livestock and Manure Management, Equation 10.3; Table 10.4.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories. Chapter 10: Emissions from Livestock and Manure Management, Equation 10.3; Table 10.4.

Wardeh, M. F. (2004). The nutrient requirements of the dromedary camel. Journal of Camel Science. 2004;1:37–45. The Camel Applied Research and Development Network (CARDN), Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).