Calculate Energy for Pregnancy

Computes the energy requirement for pregnancy (MJ/head/day) for pregnant females. This approach follows the IPCC Tier 2 partitioning framework and is applied only to female cohorts.

calc_net_energy_pregnancy(
  animal,
  cohort,
  nemain,
  parturition_rate,
  litsize,
  gest,
  idle,
  lact,
  duration,
  offtake_rate
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

cohort

Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:

FA: adult females (from age at first parturition)

FS: sub-adult females (from weaning to age at first parturition)

FJ: juvenile females (from birth to weaning)

MA: adult males (from age at first breeding)

MS: sub-adult males (from weaning to age at first breeding)

MJ: juvenile males (from birth to weaning)

nemain

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

parturition_rate

Numeric. Average annual number of parturitions per female animal (#). A herd-level reproductive performance indicator calculated as the total number of parturitions (deliveries) occurring during a year divided by the number of adult females potentially able to give birth during that year.

litsize

Numeric. Average number of offspring born per parturition (#). This value can be calculated as the total number of offspring born divided by the total number of parturitions during the year.

gest

Numeric. Duration of pregnancy period (days).

idle

Numeric. Period during which the animal is not performing any productive physiological function such as pregnancy or lactation (days).

lact

Numeric. Duration of the lactation period, defined as the number of days during which the animal is lactating (days).

duration

Numeric. Amount of time that each animal spends in a specific cohort (days).

offtake_rate

Numeric. Annual proportion of animals removed from the herd for each sex-age cohort (fraction).

Value

Numeric. Energy required for pregnancy for pregnant adult females (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

Details

Pregnancy energy is calculated only for female cohorts (i.e., FA and FS) and represents the additional energy required to support gestation.

For FA, pregnancy energy requirements are adjusted to account for the proportion of time animals spend in gestation.

For FS, only a fraction of animals is assumed to be of reproductive age; pregnancy energy requirements are therefore scaled accordingly.

For CTL and BFL — IPCC (2006, 2019):

Pregnancy energy is approximated as 10% of maintenance energy.

FA: $$ energy\_pregnancy = 0.10 \times nemain \times parturition\_rate \times \frac{gest}{365} $$

FS: $$ energy\_pregnancy = 0.10 \times nemain \times \frac{gest}{duration} \times (1 - offtake\_rate) $$

For CML — Wardeh (2004):

Pregnancy energy is estimated as 12% of maintenance energy.

FA: $$ energy\_pregnancy = 0.12 \times nemain \times parturition\_rate $$

FS: $$ energy\_pregnancy = 0.12 \times nemain \times \frac{gest}{duration} \times (1 - offtake\_rate) $$

For SHP and GTS — IPCC (2006, 2019):

Pregnancy energy is calculated as a litter-size–dependent fraction of maintenance energy.

FA: $$ energy\_pregnancy = nemain \times cpreg \times parturition\_rate \times \frac{gest}{365} $$ where \(cpreg\) is defined as:

If \(1 \le litsize \le 2\): $$ cpreg = 0.077 \times (2 - litsize) + 0.126 \times (litsize - 1) $$

If \(litsize > 2\): $$cpreg = 0.150$$

FS - Pregnancy energy is calculated using the single-birth coefficient and scaled to the proportion of reproductive individuals in the cohort. $$ energy\_pregnancy = 0.077 \times nemain \times \frac{gest}{duration} \times (1 - offtake\_rate) $$

For PGS — NRC (1998):

FA: $$ energy\_pregnancy = cgest \times litsize \times \frac{gest}{lact + gest + idle} $$

FS: $$ energy\_pregnancy = cgest \times litsize \times \frac{gest}{duration} \times (1 - offtake\_rate) $$ The default value for the gestation coefficient is \(cgest = 0.14985\) MJ piglet\(^{-1}\).

References

NRC (1998). Nutrient Requirements of Swine, 10th Revised Edition. National Academies Press, Washington, DC.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories. Chapter 10: Emissions from Livestock and Manure Management, Equation 10.13; Table 10.7.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories. Chapter 10: Emissions from Livestock and Manure Management, Equation 10.13; Table 10.7.

Wardeh, M. F. (2004). The nutrient requirements of the dromedary camel. Journal of Camel Science. 2004;1:37–45. The Camel Applied Research and Development Network (CARDN), Arab Center for the Studies of Arid Zones and Dry Lands (ACSAD).