Calculate Energy for Work

Computes the energy requirement for work (MJ/head/day), defined as the energy required for draft power. This approach follows the IPCC Tier 2 partitioning method and applies species-specific coefficients.

calc_net_energy_work(
  animal,
  cohort,
  nemain,
  work_hours_female,
  work_hours_male,
  draught_fraction_female,
  draught_fraction_male
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CHK: chickens

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

cohort

Character. Sex- and age-specific cohort code describing the production stage of the animals. Supported values include:

FA: adult females (from age at first parturition)

FS: sub-adult females (from weaning to age at first parturition)

FJ: juvenile females (from birth to weaning)

MA: adult males (from age at first breeding)

MS: sub-adult males (from weaning to age at first breeding)

MJ: juvenile males (from birth to weaning)

nemain

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

work_hours_female

Numeric. Average daily working time per adult females for CTL, BFL and CML, expressed as hours worked per head per day (hours/head/day).

work_hours_male

Numeric. Average daily working time per adult males for CTL, BFL and CML, expressed as hours worked per head per day (hours/head/day).

draught_fraction_female

Numeric. Fraction of adult females involved in draught work (fraction).

draught_fraction_male

Numeric. Fraction of adult males involved in draught work (fraction).

Value

Numeric. Energy required for work, used to estimate the energy required for draught power for CTL, BFL and CML. Assumed to be 0 for other species. (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

Details

Energy for work (energy_work) represents the additional energy required to support draught power generation by working animals.

This component is applied only to draught-capable species and is scaled by the fraction of adult animals involved in draught work (draught_fraction) and their average daily working time (work_hours).

CTL and BFL - Bamualim & Kartiarso (1985); IPCC (2006, 2019). For cattle and buffalo, draught work energy is expressed as a proportion of net energy for maintenance:

\( energy\_work = 0.1 \times NEm \times work\_hours \times draught\_fraction \)

where:

\(energy\_work\) is net energy required for maintenance (MJ head\(^{-1}\) day\(^{-1}\)),

\(0.1\) represents a 10% increase in maintenance energy per hour of work,

\(work\_hours\) is the mean number of hours worked per animal per day,

\(draught\_fraction\) is the fraction of adult males performing draught work.

CML - Wilson (1989) For camels, draught work energy is calculated using a fixed metabolizable energy cost per hour of work:

$$ energy\_work = 4 \times work\_hours \times draught\_fraction $$

where:

\(4\) is the metabolizable energy requirement for draught work (MJ head\(^{-1}\) hour\(^{-1}\)),

\(work\_hours\) is the mean daily working time per animal,

\(draught\_fraction\) is the fraction of adult males involved in draught work.

References

Bamualim A., Kartiarso (1985). Nutrition of draught animals with special reference to Indonesia. In:Draught Animal Power for Production. Australian Centre for International agricultural Research (ACIAR), Proceedings Series No. 10, ed. JW Copland. Canberra, A.C.T., Australia: ACIAR.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.11.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.11.

Wilson (1989). The nutritional requirements of camel. In: Tisserand J.-L. (ed.). Séminaire sur la digestion, la nutrition et l'alimentation du dromadaire. Zaragoza : CIHEAM. (1989). p. 171-179 (Options Méditerranéennes : Série A. Séminaires Méditerranéens; n. 2)