Calculate Ratio of Net Energy available for Maintenance in the diet (REM - Net Energy for Maintenance / Digestible Energy)

Computes the ratio of net energy available in the diet for maintenance to digestible energy.

calc_rem_maintenance(species_short, ration_digestibility_fraction = NA_real_)

Arguments

species_short

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

ration_digestibility_fraction

Numeric. Average digestibility of the feed ration, expressed as ratio of digestible to gross energy content (fraction).

Value

Numeric. Ratio of net energy available for maintenance in the diet to digestible energy consumed (fraction).

Details

This component follows the IPCC Tier 2 partitioning approach and it returns the value for ruminants (CTL, BFL, SHP, GTS) calculated as follows:

$$ net\_energy\_maintenance\_digestible\_energy\_ratio = 1.123 - 0.004092 \times (diet\_digestibility\_fraction \times 100) + 0.00001126 \times (diet\_digestibility\_fraction \times 100)^2 - \frac{25.4}{diet\_digestibility\_fraction \times 100} $$

For the Other species REM is not applicable and the function returns NA_real_.

References

Gibbs M.J., Johnson D.E. (1993) Livestock Emissions. In: International Methane Emissions. Washington, D.C., U.S.A: US Environmental Protection Agency, Climate Change Division.

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.14.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.14.

See also