Calculate Total Energy Requirement

Computes the total daily energy requirement (MJ/head/day) by summing relevant energy partitions (maintenance, activity, lactation, work, pregnancy, growth, fibre, egg deposition).

calc_total_energy_requirement(
  animal,
  nemain,
  neact,
  nelact,
  nework,
  nepreg,
  rem,
  negrow,
  nefibre,
  neegg,
  reg,
  diet_dig
)

Arguments

animal

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CHK: chickens

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

nemain

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal in equilibrium such that body energy is neither gained nor lost. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

neact

Numeric. Energy required for activity, defined as the amount of energy needed to obtain feed, water and shelter (example stall feeding, grazing large areas). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

nelact

Numeric. Energy required for lactation. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

nework

Numeric. Energy required for work, used to estimate the energy required for draft power for CTL, BFL and CML. Assumed to be 0 for other species. (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

nepreg

Numeric. Energy required for pregnancy for pregnant adult females (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

rem

Numeric. Ratio of net energy available for growth in a diet to digestible energy consumed (fraction).

negrow

Numeric. Energy required for growth (i.e., weight gain). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

nefibre

Numeric. Energy required for the synthesis of fibre for SHP, GTS and CML. Assumed to be 0 for other species. (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

neegg

Numeric. Energy required for egg deposition for CHK (MJ/head/day).

reg

Numeric. Ratio of net energy available for growth in a diet to digestible energy consumed (fraction).

diet_dig

Numeric. Average digestibility of the the feed ration, expressed as ratio of digestible to gross energy content (fraction)

Value

Numeric. Total daily energy requirement (MJ/head/day). For CTL, BFL, SHP and GTS this is expressed as gross energy intake requirement (GE). For CML and PGS the function returns the summed daily metabolizable energy requirement.

Details

The total energy requirement is computed differently depending on whether species energy requirements are expressed as net energy (NE) or metabolizable energy (ME).

Energy requirements expressed as net energy (CTL, BFL, SHP, GTS):

For these species, the calculation follows the IPCC Tier 2 structure (Equation 10.16). Net energy requirements are converted to gross energy (GE) intake using efficiency ratios for maintenance-type (\(REM\)) and growth-type (\(REG\)) functions, and diet digestibility (\(diet\_dig = DE/GE\)).

Cattle and buffalo (CTL, BFL): $$ energy\_total = \frac{\left(\frac{nemain + neact + nelact + nework + nepreg}{rem}\right) + \left(\frac{negrow}{reg}\right)}{diet\_dig} $$

Sheep and goats (SHP, GTS): Fibre production is treated as a growth-type requirement: $$ energy\_total = \frac{\left(\frac{nemain + neact + nelact + nepreg}{rem}\right) + \left(\frac{negrow + nefibre}{reg}\right)}{diet\_dig} $$

Energy requirements expressed as metabolizable energy (CML, PGS):

For these species, total energy requirement is calculated as the direct sum of the relevant daily energy components.

Camels (CML): $$ energy\_total = nemain + neact + nelact + nework + nefibre + nepreg + negrow $$

Pigs (PGS): $$ energy\_total = nemain + neact + nelact + nepreg + negrow $$

References

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.16.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.16.