Calculate Total Energy Requirement

Computes the total daily energy requirement (MJ/head/day) by summing relevant energy partitions (maintenance, activity, lactation, work, pregnancy, growth, fibre, egg deposition).

calc_total_metabolic_energy_req(
  species_short,
  metabolic_energy_req_maintenance,
  metabolic_energy_req_activity,
  metabolic_energy_req_lactation,
  metabolic_energy_req_work,
  metabolic_energy_req_pregnancy,
  net_energy_maintenance_digestible_energy_ratio,
  metabolic_energy_req_growth,
  metabolic_energy_req_fibre_production,
  metabolic_energy_req_egg_deposition,
  net_energy_growth_digestible_energy_ratio,
  ration_digestibility_fraction
)

Arguments

species_short

Character. Code identifying the livestock species. Supported values include:

PGS: pigs

CML: camels

CTL: cattle

BFL: buffalo

SHP: sheep

GTS: goats

metabolic_energy_req_maintenance

Numeric. Energy required for maintenance, defined as the amount of energy needed to keep the animal at equilibrium such that body energy is neither gained nor lost (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

metabolic_energy_req_activity

Numeric. Energy required for activity, defined as the amount of energy needed to support animal movement and physical activity (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

metabolic_energy_req_lactation

Numeric. Energy required for lactation (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

metabolic_energy_req_work

Numeric. Energy required for work, used to estimate the energy required for draught power for CTL, BFL and CML (MJ/head/day). Assumed to be 0 for other species. Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

metabolic_energy_req_pregnancy

Numeric. Energy required for pregnancy for pregnant females (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

net_energy_maintenance_digestible_energy_ratio

Ratio of net energy available for maintenance in the diet to digestible energy consumed (fraction).

metabolic_energy_req_growth

Numeric. Energy required for growth (i.e., weight gain) (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS.

metabolic_energy_req_fibre_production

Numeric. Energy required for the synthesis of fibre for SHP, GTS and CML. Assumed to be 0 for other species (MJ/head/day). Expressed as net energy for CTL, BFL, SHP, GTS and as metabolizable energy for CML and PGS (MJ/head/day).

metabolic_energy_req_egg_deposition

Numeric. Net energy for egg production (MJ/head/day).

net_energy_growth_digestible_energy_ratio

Numeric. Ratio of net energy available for growth in the diet to digestible energy consumed (fraction)

ration_digestibility_fraction

Numeric. Average digestibility of the the feed ration, expressed as ratio of digestible to gross energy content (fraction).

Value

Numeric. Total daily energy requirement (MJ/head/day). For CTL, BFL, SHP and GTS this is expressed as gross energy intake requirement (GE). For CML and PGS the function returns the summed daily metabolizable energy requirement.

Details

This component follows the IPCC Tier 2 partitioning approach and the calculation is computed differently depending on whether species energy requirements are expressed as net or metabolizable energy.

Species-specific approach:

Energy requirements expressed as net energy (CTL, BFL, SHP, GTS)

CTL and BFL: $$ energy\_requirement\_total = \frac{ \left( \frac{ energy\_requirement\_maintenance + energy\_requirement\_activity + energy\_requirement\_lactation + energy\_requirement\_work + energy\_requirement\_pregnancy }{REM} \right) + \left( \frac{energy\_requirement\_growth}{REG} \right) }{diet\_digestibility\_fraction} $$

SHP and GTS: $$ energy\_requirement\_total = \frac{ \left( \frac{ energy\_requirement\_maintenance + energy\_requirement\_activity + energy\_requirement\_lactation + energy\_requirement\_pregnancy }{REM} \right) + \left( \frac{energy\_requirement\_growth + energy\_requirement\_fibre }{REG} \right) }{diet\_digestibility\_fraction} $$

Energy requirements expressed as metabolizable energy (CML, PGS)

For these species, the total daily requirement is computed as the direct sum of relevant energy components (MJ/head/day).

CML: $$ energy\_requirement\_total = energy\_requirement\_maintenance + energy\_requirement\_activity + energy\_requirement\_lactation + energy\_requirement\_work + energy\_requirement\_fibre\_production + energy\_requirement\_pregnancy + energy\_requirement\_growth $$

PGS: $$ energy\_requirement\_total = energy\_requirement\_maintenance + energy\_requirement\_activity + energy\_requirement\_lactation + energy\_requirement\_pregnancy + energy\_requirement\_growth $$

References

IPCC. (2019). 2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.16.

IPCC. (2006). 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management, Equation 10.16.

See also