gleam: Global Livestock Environmental Assessment Model (GLEAM-X)

GLEAM-X is the official R implementation of FAO’s Global Livestock Environmental Assessment Model (GLEAM). The package provides a modular, transparent framework for simulating livestock production systems and quantifying their environmental impacts. Methodological background: MacLeod et al. (2017) "Invited review: A position on the Global Livestock Environmental Assessment Model (GLEAM)" doi:10.1017/S1751731117001847 . Further information: https://www.fao.org/gleam/en/.

See also

Author

Maintainer: Ahmed Jou ahmed@applitics.fr (ORCID) (https://applitics.fr)

Authors:

Yassine Elaouni yassine@applitics.fr (https://applitics.fr)

Dominik Wisser dominik.wisser@fao.org (ORCID) (GLEAM team lead, FAO)

Lydia Lanzoni lydia.lanzoni@fao.org (ORCID) (Livestock and climate change specialist, FAO)

Giuseppe Tempio giuseppe.tempio@fao.org (ORCID) (Livestock environmental impact analyst, FAO)

Other contributors:

Phyllis Ndungu' Phyllis.Ndungu@fao.org (Livestock and climate change specialist, FAO) [contributor]

Yushan Li Yushan.Li@fao.org (Livestock and sustainability specialist, FAO) [contributor]

Manling Xu Manling.Xu@fao.org (Livestock data intern, FAO) [contributor]

Zixin Wang Zixin.Wang@fao.org (Livestock data intern, FAO) [contributor]

Narindra Rakotovao Narindra.Rakotovao@fao.org (Livestock policy officer, FAO) [contributor]

Stewart Chikoloma Stewart.Chikomola@fao.org (Livestock data specialist, FAO) [contributor]

David OBrien david@applitics.fr (ORCID) (https://applitics.fr) [contributor]

Food and Agriculture Organization of the United Nations FAO (Original GLEAM model framework) [copyright holder]

Federal Ministry of Agriculture, Food and Regional Identity of Germany BMELH (Funding support) [funder]